Split config.go into files
